Store Support/Projects/TDA Insights/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-bigquery==3.14.0
python-pptx==0.6.21
pandas==2.1.0
gunicorn==21.2.0
